app/requirements.txt
flask
flask_sqlalchemy
flask-login
Flask-Migrate
werkzeug
psycopg2-binary
grpcio
grpcio-tools
grpcio-reflection
aiokafka
